dataIO: fixed bug in PNG 2 H5

+++ utilities/dataIO.py
@@ -166,7 +166,7 @@ def PNG2H5(directory, filename, dataset, dtype=np.int32):
 
         # add this element
         if len(im.shape) == 3 and dtype == np.int32: h5output[iz,:,:] = 65536 * im[:,:,0] + 256 * im[:,:,1] + im[:,:,2]
-        elif len(im.shape) == 3 and dtype == np.uint8: h5output[iz,:,:] = (im[:,:,0] + im[:,:,1] + im[:,:,2]) / 3
+        elif len(im.shape) == 3 and dtype == np.uint8: h5output[iz,:,:] = ((im[:,:,0].astype(np.uint16) + im[:,:,1].astype(np.uint16) + im[:,:,2].astype(np.uint16)) / 3).astype(np.uint8)
         else: h5output[iz,:,:] = im[:,:]
 
     WriteH5File(h5output, filename, dataset)
